fixed bug when ws server already exists

+++ gym_anm/envs/anm6_env/rendering/py/rendering.py
@@ -1,6 +1,7 @@
 import webbrowser
 import json
 import os
+import websocket
 
 from websocket import create_connection
 import time
@@ -57,8 +58,9 @@ def start(title, dev_type, p_max, q_max, s_rate, v_magn_min, v_magn_max, soc_max
     p = os.path.join(http_server.address, RENDERING_RELATIVE_PATH)
     webbrowser.open_new_tab(p)
 
-    time.sleep(3)  # To make sure the WS server has started.
+    # time.sleep(3)  # To make sure the WS server has started.
     ws = create_connection(ws_server.address)
+
     message = json.dumps({'messageLabel': 'init',
                           'deviceType': dev_type,
                           'pMax': p_max,
@@ -147,7 +149,9 @@ def close(http_server, ws_server):
     """
 
     http_server.process.terminate()
-    ws_server.process.terminate()
+
+    if ws_server is not None:
+        ws_server.process.terminate()
 
 
 def write_html():
